% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4prot.R
\name{getPackagesInBiocView}
\alias{getPackagesInBiocView}
\title{Packages in a biocView}
\usage{
getPackagesInBiocView(view, rep = c("BioCsoft", "BioCann", "BioCexp",
  "BioCextra"), biocVersion)
}
\arguments{
\item{view}{The biocView of interest. For example
\code{"Proteomics"}.}

\item{rep}{Repository of interest. One of \code{"BioCsoft"},
\code{"BioCann"}, \code{"BioCexp"} or \code{"BioCextra"}.}

\item{biocVersion}{A \code{character} with the Bioconductor
version of interest. For example \code{"2.14"}.}
}
\value{
An instance of class \code{BiocView}. \code{NULL} if the
the biocView was not found.
}
\description{
Finds the package names that have a specific biocView.
}
\author{
Laurent Gatto
}
