\name{id}
\alias{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzid}
\docType{data}

\title{
  An \code{mzIdentML} file
}

\description{

  This file has been generated by searching the raw \code{mzXML} file of
  the ProteomeXchange \code{PXD000001} data set against the
  \code{erwinia_carotovora.fasta} using the MSGF+ search engine:
  \code{java -jar ~/bin/MSGFPlus.20140630/MSGFPlus.jar -s
  TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML -d
  erwinia_carotovora.fasta -inst 1 -protocol 2}
  
}

\examples{
## source files to repeat the search
library("rpx")
px <- PXDataset("PXD000001")
pxfiles(px)

f <- dir(system.file("mzid", package = "RforProteomics"), full.names=TRUE)
library("mzID")
id <- mzID(f)
id
}

\keyword{datasets}
