% Generated by roxygen2 (4.0.1): do not edit by hand
\name{packageDF}
\alias{packageDF}
\title{Package descriptions}
\usage{
packageDF(x, nsub = TRUE, version = TRUE)
}
\arguments{
\item{x}{An instance of class \code{BiocView}, as produced by
\code{getPackagesInBiocView}.}

\item{nsub}{A \code{logical} indicating \code{"\n"} are to be
replaced by a space.}

\item{version}{A \code{logical} specifying if the package version
should be added.}
}
\value{
A \code{data.frame} with package information.
}
\description{
Format a \code{BiocView} as a \code{data.frame}.
}
\author{
Laurent Gatto
}

