% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProtViz}
\alias{ProtViz}
\title{Opens the visualisation vignette}
\usage{
ProtViz()
}
\value{
An instance of class \code{vignette}. Used for its side
effect, opening the vignette.
}
\description{
Opens the visualisation vignette
}
\author{
Laurent Gatto
}

