% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RforProteomics}
\alias{RforProteomics}
\title{Opens RforProteomics vignettes}
\usage{
RforProteomics()
}
\value{
An instance of class \code{vignette},
as returned by \code{\link{vignette}}
}
\description{
Opens the package vignettes.
}
\author{
Laurent Gatto
}

