% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r4prot.R
\name{RProtVis}
\alias{RProtVis}
\alias{RProtViz}
\title{Opens the visualisation vignette}
\usage{
RProtVis()
}
\value{
An instance of class \code{vignette}. Used for its side
effect, opening the vignette.
}
\description{
Opens the visualisation vignette
}
\author{
Laurent Gatto
}

