% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PXD000001.R
\name{getPXD000001mzTab}
\alias{getPXD000001mzTab}
\title{Download the PXD000001 mzTab file}
\usage{
getPXD000001mzTab(destdir = ".")
}
\arguments{
\item{destdir}{A \code{character} with the destination folder.}
}
\value{
Invisibly returns the name of the downloaded file.
}
\description{
Unless already present, downloads the PXD000001 mzTab file
in the \code{destdir} directory. The resulting file is named
\code{F063721.dat-mztab.txt}.
}
\author{
Laurent Gatto
}

