\name{getPXD000001mzXML}
\alias{getPXD000001mzXML}
\title{Download the PXD000001 mzXML file}
\usage{
  getPXD000001mzXML(destdir = ".")
}
\arguments{
  \item{destdir}{A \code{character} with the destination
  folder.}
}
\value{
  Invisibly returns the name of the downloaded file.
}
\description{
  Unless already present, downloads the PXD000001 mzXML
  file in the \code{destdir} directory. The resulting file
  is named
  \code{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML}.
}
\author{
  Laurent Gatto
}

