% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4prot.R
\name{RforProteomics}
\alias{RforProteomics}
\title{Opens the RforProteomics vignette}
\usage{
RforProteomics()
}
\value{
An instance of class \code{vignette}. Used for its side
effect, opening the vignette.
}
\description{
Opens the package vignettes.
}
\author{
Laurent Gatto
}

