% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4prot.R
\name{RforProteomics}
\alias{RforProteomics}
\alias{RProtVis}
\title{Opens the package vignettes}
\usage{
RforProteomics()

RProtVis()
}
\value{
An instance of class \code{vignette}. Used for its side effect,
opening a vignette.
}
\description{
Opens the package and visualisation vignettes.
}
\author{
Laurent Gatto
}
